+++ lib/src/exception.dart
@@ -25,13 +25,30 @@ final class FalsifiedException<T> implements Exception {
     required this.example,
     this.description,
     this.seed,
+    this.exception,
+    this.stackTrace,
   });
 
   final T example;
   final String? description;
   final int? seed;
+  final Exception? exception;
+  final StackTrace? stackTrace;
 
   @override
-  String toString() =>
-      'Falsifying example: ${description ?? example}${seed != null ? ' (seed $seed)' : ''}';
+  String toString() {
+    final buffer = StringBuffer()
+      ..writeln(
+          'Falsifying example: ${description ?? example}${seed != null ? ' (seed $seed)' : ''}');
+
+    if (exception != null) {
+      buffer.writeln('Exception: $exception');
+    }
+    if (stackTrace != null) {
+      buffer
+        ..writeln('Stack trace:')
+        ..writeln(stackTrace);
+    }
+    return buffer.toString();
+  }
 }